requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Werkzeug==2.3.7
SQLAlchemy==2.0.36
WTForms==3.1.0
Flask-WTF==1.1.1
yfinance==0.2.65
pandas==2.3.1
numpy==2.2.1
requests==2.32.4
python-dotenv==1.0.0
finance-datareader==0.9.96
pytz==2025.2
ta==0.10.2
matplotlib==3.9.2
mplfinance==0.12.10b0
google-generativeai==0.3.2
psycopg2-binary==2.9.10
concurrent-log-handler==0.9.24
sendgrid==6.11.0
premailer==3.10.0
cryptography==42.0.8
APScheduler==3.10.4